# server/docs/swagger.yaml
definitions:
...
  github_com_flipped-aurora_gin-vue-admin_server_plugin_limitedActivity_model_request.MiniOrderCreate:
    properties:
      activityId:
        type: integer
      contactName:
        type: string
      contactPhone:
        type: string
      quantity:
        minimum: 1
        type: integer
    required:
    - activityId
    - contactName
    - contactPhone
    - quantity
    type: object
...
  model.Activity:
    properties:
      ID:
        description: 主键ID
        type: integer
      canSignup:
        description: 虚拟字段：当前是否可报名（显示中且在时间窗内且有余量）
        type: boolean
      coverImage:
        type: string
      createdAt:
        description: 创建时间
        type: string
      createdBy:
        type: integer
      detail:
        type: string
      endTime:
        type: string
      groupQr:
        type: string
      marketPrice:
        type: number
      name:
        type: string
      price:
        type: number
      quota:
        type: integer
      remaining:
        description: 虚拟字段：剩余可报名名额
        type: integer
      serviceQr:
        type: string
      sold:
        type: integer
      startTime:
        type: string
      status:
        type: integer
      updatedAt:
        description: 更新时间
        type: string
      updatedBy:
        type: integer
    type: object
...
  response.PageResult:
    properties:
      list: {}
      page:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
    type: object
...
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
info:
  contact: {}
  description: 使用gin+vue进行极速开发的全栈开发基础平台
  title: Gin-Vue-Admin Swagger API接口文档
  version: v2.8.9
paths:
...
  /limitedActivity/activity/createActivity:
    post:
      consumes:
      - application/json
      parameters:
      - description: 活动信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.Activity'
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 创建限时活动
      tags:
      - LimitedActivity
  /limitedActivity/activity/deleteActivity:
    delete:
      parameters:
      - description: 活动ID
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 删除限时活动
      tags:
      - LimitedActivity
  /limitedActivity/activity/deleteActivityByIds:
    delete:
      consumes:
      - application/json
      parameters:
      - description: ID列表
        in: body
        name: data
        required: true
        schema:
          items:
            type: integer
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 批量删除限时活动
      tags:
      - LimitedActivity
  /limitedActivity/activity/findActivity:
    get:
      parameters:
      - description: 活动ID
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 查询成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.Activity'
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 查询限时活动
      tags:
      - LimitedActivity
  /limitedActivity/activity/getActivityList:
    get:
      parameters:
      - description: 关键字
        in: query
        name: keyword
        type: string
      - in: query
        name: name
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页大小
        in: query
        name: pageSize
        type: integer
      - in: query
        name: status
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.PageResult'
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 限时活动列表
      tags:
      - LimitedActivity
  /limitedActivity/activity/updateActivity:
    put:
      consumes:
      - application/json
      parameters:
      - description: 活动信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.Activity'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 更新限时活动
      tags:
      - LimitedActivity
...
  /limitedActivity/mini/order/create:
    post:
      consumes:
      - application/json
      description: 提交人次、联系人姓名与手机号创建待支付订单并占用名额。支付请调用 POST /mini/pay/create，body 传 {"orderType":"limitedActivity","orderId":订单ID}
      parameters:
      - description: 小程序登录 token
        in: header
        name: x-token
        required: true
        type: string
      - description: 报名信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/github_com_flipped-aurora_gin-vue-admin_server_plugin_limitedActivity_model_request.MiniOrderCreate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  type: object
                msg:
                  type: string
              type: object
      summary: 报名创建订单
      tags:
      - 小程序-限时活动
  /limitedActivity/mini/order/detail:
    get:
      consumes:
      - application/json
      description: 含核销进度、群二维码、客服二维码；用户不可自助退款
      parameters:
      - description: 小程序登录 token
        in: header
        name: x-token
        required: true
        type: string
      - description: 订单ID
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  type: object
                msg:
                  type: string
              type: object
      summary: 活动订单详情
      tags:
      - 小程序-限时活动
  /limitedActivity/mini/order/myList:
    get:
      consumes:
      - application/json
      description: 'orderType: pending_payment|pending_verify|completed；不传返回全部'
      parameters:
      - description: 小程序登录 token
        in: header
        name: x-token
        required: true
        type: string
      - description: pending_payment|pending_verify|completed
        in: query
        name: orderType
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页条数
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.PageResult'
                msg:
                  type: string
              type: object
      summary: 我的活动订单列表
      tags:
      - 小程序-限时活动
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: x-token
    type: apiKey
swagger: "2.0"
...